// src/app/model/supplier.ts
export interface Supplier {
    supplier_id: string;
    supplier_hash_id: string;
    supplier_name: string;
    supplier_tel: string;
    supplier_email: string;
    supplier_company: string;
    supplier_tax_number: string;
    supplier_info: string;
    created_at: Date;
    updated_at: Date;
    supplier_address: string;
}